fix: login logout & user detail

+++ app/Services/User/UserServiceImplement.php
@@ -2,13 +2,14 @@
 
 namespace App\Services\User;
 
-use LaravelEasyRepository\Service;
+use App\Helpers\Strings;
 use App\Repositories\User\UserRepository;
-use App\Traits\PrintLog;
-use Illuminate\Support\Facades\{Auth, Hash, Session, Validator};
+use App\Services\BaseService;
+use Illuminate\Contracts\Encryption\DecryptException;
+use Illuminate\Support\Facades\{Crypt, Hash, Session};
 
-class UserServiceImplement extends Service implements UserService{
-    use PrintLog;
+class UserServiceImplement extends BaseService implements UserService{
+
     /**
      * don't change $this->mainRepository variable name
      * because used in extends service class
@@ -45,17 +46,23 @@ class UserServiceImplement extends Service implements UserService{
                 $data = [
                     'access_token'  => $tokenPlain,
                     'token_type'    => 'Bearer',
-                    'user'          => $user,
+                    'user'          => $user->toArray(),
                 ];
+                $data['user']['id'] = Crypt::encrypt($data['user']['id']);
+                $data['user']['name'] = $user->personal->name;
+                $data['user']['simple_name'] = Strings::simpleString($user->personal->name);
+                unset($data['user']['personal_id']);
+                unset($data['user']['personal']);
 
                 $session = [
-                    'id'         => $user->id,
+                    'id'         => Crypt::encrypt($user->id),
                     'avatar'     => $user->avatar,
                     'email'      => $user->email,
                     'name'       => $user->name,
-                    'branch_id'  => $user->branch_id,
-                    'company_id' => $user->company_id,
+                    'branch_id'  => Crypt::encrypt($user->branch_id),
+                    'company_id' => Crypt::encrypt($user->company_id),
                     'token'      => $tokenPlain,
+                    'token_id'   => Crypt::encrypt($tokenId),
                 ];
                 session($session);
 
@@ -65,6 +72,7 @@ class UserServiceImplement extends Service implements UserService{
             }
         } else {
             $result['message']  = __('auth.account_not_found');
+            $result['code']     = 404;
         }
 
         return $result;
@@ -72,17 +80,64 @@ class UserServiceImplement extends Service implements UserService{
 
     function signout($request): array
     {
-        $userId  = Session::get('id');
-
-        $user = $this->mainRepository->getUserById($userId)->first();
-        $user->tokens()->delete(); // remove token
-        $request->session()->flush(); // remove session
         $result = [
-            'success'   => true,
-            'code'      => 200,
-            'message'   => __('auth.logout'),
+            'success'   => false,
+            'code'      => 400,
         ];
+        try {
+            $encryptUserId  = Session::get('id');
+            if(!empty($encryptUserId)){
+                $userId = Crypt::decrypt($encryptUserId);
+                $user = $this->mainRepository->getUserById($userId);
+                if($user){
+                    $user->tokens()->delete(); // remove token
+                }
+            }
+            $request->session()->flush(); // remove session
+            $result = [
+                'success'   => true,
+                'code'      => 200,
+                'message'   => __('auth.logout'),
+            ];
+        } catch(DecryptException $e){
+            $result['message'] = __('content.payload_invalid');
+        } catch (\Throwable $th) {
+            if(config('app.debug') == true){
+                $result['message'] = $th->getMessage();
+            } else {
+                $result['message'] = __('content.something_error');
+            }
+        }
+        return $result;
+    }
 
+    public function getById($encryptId)
+    {
+        $result = [
+            'success'   => false,
+            'code'      => 400,
+        ];
+        try {
+            $id = Crypt::decrypt($encryptId);
+            $user = $this->mainRepository->getUserById($id);
+            $userArray = $user->toArray();
+            unset($userArray['id']);
+
+            $result = [
+                'success'   => true,
+                'code'      => 200,
+                'message'   => __('content.ok'),
+                'data'      => $userArray,
+            ];
+        } catch(DecryptException $e){
+            $result['message'] = __('content.payload_invalid');
+        } catch (\Throwable $th) {
+            if(config('app.debug') == true){
+                $result['message'] = $th->getMessage();
+            } else {
+                $result['message'] = __('content.something_error');
+            }
+        }
         return $result;
     }
 }
